refactor: move ExtensionBasedMapper into mapping directory
